refactor(web): clean up imports and remove unused components
- Removed unused imports from various components, including  from profile and  from dashboard.
- Streamlined import statements in settings and daily plan components for better organization.
- Enhanced code maintainability by eliminating obsolete components and ensuring only necessary imports are retained.
